Add http_port range validation (0 or 1-65535)
Change http_port type from u16 to u32 to allow out-of-range values to be
caught by validation. Add HttpPortInvalid error variant and validation check
for http_port > 65535. Add test for http_port=65536 producing HttpPortInvalid.
http_port=0 (disabled) remains valid per existing test.

+++ src/config/static_config.rs
@@ -33,7 +33,7 @@ pub fn default_shutdown_timeout_secs() -> u64 {
 pub struct ListenerConfig {
     pub bind_addr: String,
     #[serde(default = "default_http_port")]
-    pub http_port: u16,
+    pub http_port: u32,
     #[serde(default = "default_https_port")]
     pub https_port: u16,
     pub tls: TlsConfig,
@@ -42,7 +42,7 @@ pub struct ListenerConfig {
 }
 
 #[allow(dead_code)]
-fn default_http_port() -> u16 {
+fn default_http_port() -> u32 {
     80
 }
